Rename pages, move detail page components
