+++ src/components/DetailPage/Card.tsx
