+++ src/pages/bookmarks.tsx
@@ -4,7 +4,7 @@ import LeftSidebar from '@/components/LeftSidebar'
 
 const inter = Inter({ subsets: ['latin'] })
 
-const savedProjects = () => (
+const Bookmarks = () => (
   <main className={`min-h-screen p-24 ${inter.className} bg-gray-900 text-gray-100`}>
     <div className="flex flex-col">
       <LeftSidebar />
@@ -14,4 +14,4 @@ const savedProjects = () => (
     </div>
   </main>
 )
-export default savedProjects
+export default Bookmarks
