+++ src/pages/docs.tsx
@@ -3,7 +3,7 @@ import LeftSidebar from '@/components/LeftSidebar'
 
 const inter = Inter({ subsets: ['latin'] })
 
-const Documentation = () => (
+const Docs = () => (
   <main className={`min-h-screen p-24 ${inter.className} bg-gray-900 text-gray-100`}>
     <div className="flex flex-col">
       <LeftSidebar />
@@ -11,4 +11,4 @@ const Documentation = () => (
     </div>
   </main>
 )
-export default Documentation
+export default Docs
